requirements.txt
asgiref==3.5.2
blinker==1.5
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Mail==0.9.1
Flask-WTF==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
sqlparse==0.4.3
tzdata==2022.5
Werkzeug==2.2.2
